fix runtime error

+++ Src/HardwareContext.h
@@ -86,5 +86,5 @@ struct HardwareContext
 
 extern HardwareContext globalHardwareContext;
 
-extern std::unordered_map<uint32_t, ResourceManager::ImageHardwareWrap> imageGlobalPool;
-extern std::unordered_map<uint32_t, ResourceManager::BufferHardwareWrap> bufferGlobalPool;
+extern std::unordered_map<uint64_t, ResourceManager::ImageHardwareWrap> imageGlobalPool;
+extern std::unordered_map<uint64_t, ResourceManager::BufferHardwareWrap> bufferGlobalPool;
